PP-5255 Introduce EventDigest
The EventDigest is an object that represents the
summary of a list of events pertinent to the further
processing of those events. Basically that amounts to
the timestamp and event type of the latest event by
timestamp, as well as a compressed digest of the
event details attached to each event. The idea is that
this EventDigest is easily convertible into a
Transaction.