PP-5255 Insert transaction on recieving event
Adds the connecting logic that takes an EventDigest and creates and
inserts a transaction.

At the moment this simply inserts a transaction, so will not handle multiple
events. It also hardcodes the state of the transaction, so the mapping
logic needs writing.

In order to do this I have dropped some constraints on the transaction table.
This is because we don;t know what order we are going to recieve events,
so we may have only a minimal amount of information to create the
transaction, therefore we cannot enforce non-null constraints on things like
description, reference etc.

Another mildly controversial thing I have done is equate the event
`resourceExternalId` with the transaction externalId. I think this
is correct, as ultimately we need to be able to retrieve things from
connector and ledger with the same id.

This commit also does not contain any db transactions around
handling the event and creating transaction. I have thought about
this, but will hold off worrying about it for another commit.