PP-5357 Only use enum for salient events
Removes EventType enum in favour of SalientEventType.
There is now no validation of event type on reception
of event, so emitters can send any event type they
want. The SalientEventType enum will be used to
translated a known set of events to a known set of
transaction states. This is done via the EventDigest,
which now stores the 'latestSalientEventType'.