PP-5462 Store and display capture_submitted_date
Use capture_submitted_time consistently throughout date.
This is an attempt to be consistent - we have used _date everywhere
else for timestamps, and this lines up with captured_date

This commit ensures capture_submitted_date  persisted and retrieved
correctly, using a pact test to test the end to end behaviour.