PP-5407: allow event digest with 0 salient events
If no salient events are included in a digest we currently throw an
exception and reject the event message (sending it back to SQS), this is
somewhat against the Ledger "process events in any order" idea. Instead
map no salient events to an `UNDEFINED` type to be overridden when
salient events show up.