PP-5462 Move capture_submitted_date to tx details
Move capture submitted date to `transaction_details`

As part of this I have changed how transaction_details are created. Instead of mapping to an object (which means we have to specify each
field we wish to remain in transaction details) we now just write the whole thing in. This has the advantage of not having to update ledger each time a field that we do not consider to be a top level column is added to an event somewhere.