PP-5658: report on card brand label to match Connector
Connector responds with with the card brand human readable labels found
in Connector's database. This is now passed through the payment details
entered event and stored in the `card_brand_label` attribute of
transaction details. Send this in favour of the card brand key (which
can still be searched/ filtered as normal).