PP-5670: Specify the consumer when running contract test
Currently when a consumer build is kicked off, the provider contract test will
run in its entirety. For example, in a selfservice build, the ledger provider
contract tests will run as well. However, what is run is:

publicapi->ledger
selfservice->ledger
connector->ledger

We really only want to run selfservice->ledger in this case.

This commit will allow a dynamic creation of the test suite to run based on a
CONSUMER system property. This will mean shorter run times when a provider
contract test is run as part of a specific consumer build.