ZD3867367: Include ingest lag on event log lines
We currently have no visibility on the lag time for when an event gets emitted
and when it is processed by ledger. This makes it difficult for us to deal with
the query relating to this zendesk ticket about certain transactions not
appearing in calls to `/v1/payments?state=success&from_date={fromdate}` for
example.