PP-6076 Include MOTO column in CSV
If the GET /v1/transaction endpoint is called with Accept header
"text/csv" and query parameter `moto_header=true` then include a MOTO
column with a boolean type as the final column before the metadata
column.

If the query parameter is not provided or false then do not include the
column.

Add some additional tests to TransactionResourceIT to test for the
including/excluding the optional columns, including the "Net" and "Fee"
columns. Inline a private method in this test class that was only used
in one place.