PP-6310 Fix pact bug
Ledger is the only Pay microservice that is both a
consumer (of connector) and a provider (to publicapi
and selfservice). This led to a weird bug in the
Ledger pact tests. In order to run a pact test you
need to specify the `CONSUMER_TAG`. When you run a
consumer pact test, this is used to tag the pact that
the test produces. When you run a provider test, it
specifies the tag to use when retrieving pacts from a
consumer to be validated. The pact tests were run
against Ledger with the command
```
mvn clean package pact:publish -DrunContractTests=true -DPACT_CONSUMER_TAG=${branchName} …
```
This command ran both the set of tests where Ledger is
a provider and those where Ledger is a consumer. For
the provider tests it fetched the pacts from the
consumer tagged with the value of ${branchName}.
However, ${branchName} was set to be the value of the
name (actually not the name, but an identifier from
Jenkins like `PR-691`) of the current Ledger branch.
All of this fed into the failure of this build
https://build.ci.pymnt.uk/job/pay-ledger/job/PR-691/3/execution/node/32/log/
The build asks for the publicapi pact tagged with
`PP-691` to be validated against the current version of
ledger. Because `PR-691` is the Ledger "branch name",
this is obviously wrong. Unfortunately there was a pact
tagged `PP-691` from 7 months ago. Ledger no longer
satisfied this pact, so the build failed.

To fix this problem, I have separated pact tests where
Ledger is the consumer from those where it is the
provider, and run them separately, ensuring the correct
consumer tag is being set in each case.

To do this I have introduced a new maven profile,
activated by the system property
`runConsumerContractTests`. I have modified the
existing profile activated by `runContractTests`, to
only run the contract tests where Ledger is the
provider. I have not changed `runContractTests` to
`runProviderContractTests` because other builds invoke
Ledger's provider contract tests using this flag. I
have also made a small change to the default profile to
set it to `activeByDefault` which makes it play nicely
with the other flags, and also seems more intuitive.