PP-6445 Use env var for Pact broker url
We are using a diffferent pact broker for ci running on concourse so
make the pact broker url configurable from an env var `PACT_BROKER_HOST`.
Default value is the existing broker used for Jenkins CI.

Note that we cannot use the env var `PACT_BROKER_URL` which is used for
the maven pact plugin to publish pacts since that requires a url and not
just the host.