PP-6504 Explicitly set priority of content type
In the case where the client does not specify Accepts header,
we want ledger to respond with application/json. At the moment
the response type will be indeterminate if no Accepts header is
set and is determined by version of JDK.
Use qs factor to explicitly indicate that `application/json`
should be the preferred content type.
See http://www.apacheweek.com/features/negotiation