PP-6657: deprecate refund reference field in favour transaction id
Connector has always provided the gateway transaction id through the
reference field. This behaviour is confusing on both the provider and
consumer.

Now that the column in connector has been updated to
`gateway_transaction_id`, reflect that in the `Refund` model.

Description is also removed as that will never belong to a refund.