PP-6850 Re-project refunds when payment updated
When a payment event is received, if it has associated refund,
re-calculate the TransactionEntity for the refunds and persist.

This is to ensure that any payment details that we store on the refund
are kept in sync with the payment.

Avoid database lookups to get refunds when we know we won't have to do
any updates to refunds for the following cases:
- The payment has no previous salient events which map to the "SUCCESS"
state, as in this case there cannot be refunds for the payment.
- The event we are processing has no "event_data" associated with it, as
there will be no payment details that need to be updated on the refund.
- The event has type "REFUND_AVAILABILITY_UPDATED" as these events
contain "event_data" but we know that we don't need to update anything
on the refund in this case.