BAU Revert Junit Jupiter bump to 5.7
This reverts commit 6788e23ab4fe165b667b6d4affbeb0d49823c466.

This version bump meant that running the consumer contract tests no longer
resulted in the pacts being created in `target/pacts` or AFAIK anywhere.
This meant they could not be pushed to the pact broker and deploying to
environments failed since CI was unable to tag a version of Ledger which
the pact broker had no knowledge of.