PP-7696 Temporarily disable database source enum test
Temporarily disable
TransactionDaoIT.sourceTypeInDatabase_shouldMatchValuesInEnum,
which checks that the values of the source enum in the database
match those of the Source enum in the Java code.

We’re adding a new CARD_AGENT_INITIATED_MOTO value and since our
database migrations are decoupled from our code changes, we can’t
update both ledger’s database and the imported enum atomically.

So we’re going to have to disable the test, migrate the database,
update the Java enum, then re-enable the test.