PP-7832 add try-catch for race condition
we've discovered that inserting new keys to `metadata_key` table might
trigger a race condition whereupon two payments with same metadata key
can try to insert the keys. Thus one of them is going to throw an error.
This solution is not the most elegant but it works. It lets the
insertion of transaction_metadata be triggered and skipping it.

- log exception if race condition happens

with @stephencdaly