PP-7906 Reproject transaction_metadata
Reproject `transaction_metadata` rows from the EventDigest when an
event is processed with the field `reproject_domain_object` set to true.

This will not take into account any of the event details on the event
being processed, only previous events for the transaction from the
database.

When an event is processed without this field, or set to false, continue
to only upsert to the `transaction_metadata` when the event being
processed has `external_metatada` in its event details.