PP-9017 Publish incoming events to SNS
Adds a new action for ledger on recieving an event.
If SNS is enabled, Ledger publishes the event to AWS SNS,
using the provide topic.

If SNS is not enabled then publish should not be tried
and everything should behave as before.

I have introduced an enum to represent available SNS topics
to rest of app - feels nicer and safer than any of the
alternatives.

The publisher itself is very simple - don't think anything
controversial.

At the moment if any exception in SNS publishing will
be swallowed and only logged. This is to make totally sure
this does not break environments where SNS is not currently
provisioned.

Set up emitter and hook in to event processing life
cycle