PP-8772: limit search queries to configured search timeout
Specify `setQueryTimeout` on search specific DAO calls for transactions.
This will instruct JDBI to configure the JDBC driver to cancel and clean
up the database process after a specified number of seconds.

This has been tested locally to ensure that the timeout correctly ends
the postgres process ID and throws the handled exception.

Default timeout to 55 seconds which will be a safe maximum before an
egress/ proxy TCP timeout. This can be configured by the environment.