PP-9531: agreement and payment instrument DAO
DAO and JDBI row mappers for agreements and payment instruments.

Fetch agreements with their latest payment instrument using Postgres
`SELECT DISTINCT ON`.

Test coverage to assert that row mappers are correctly mapping and
casting fields.

Test coverage to assert that the latest payment instrument (by created
date) is associated with an agreement when pulled out of the database
and mapped to an agreement entity.