PP-9626 Project disputes based on environment variables
Decide whether to project disputes based on environment variables:
- PROJECT_LIVE_PAYMENTS_DISPUTE_EVENTS_FROM_DATE for live transactions
- PROJECT_TEST_PAYMENTS_DISPUTE_EVENTS_FROM_DATE for test transactions

This is so that disputes aren't returned to selfservice in transactions
searches before we are ready to return these to end users. We have
decided to not project until this date, rather than filtering out
disputes from the transaction search queries as the latter may have
performance implications.

Currently, if a dispute exists for an account and we attempt to search
for transactions that will include disputes, an error will be returned.
This will be fixed when we have added in mapping for dispute responses,
but should not cause issues for now as projection will be disabled on
all environments until this is implemented.