PP-9545: support X-Consistent param on get agreement resource
Adds branching logic to the get agreement resource to optionally check
if the current event stream has more events than the most recently
projected resource.

If `X-Consistent` is set to true, fetch the event digest for that
resource. If there are more events in the event digest than the
projection snapshots event count, return a new projection for the event
digest, otherwise return the existing projection.

If `X-Consistent` is set to false or the header has not been set,
directly return the existing projection.