PP-9671 Accept transaction_type query parameter
Accept a transaction_type query parameter for the
`/v1/transaction/{parentTransactionExternalId}/transaction` endpoint.

This will filter the child transactions for the specified parent
transaction to only return results with the requested transaction type.

This will be used by connector to query for refunds for a transaction
without also returning disputes.

Regenerate the `.secrets.baseline` using v1.2.0 of detect-secrets, which
we are now using in pay-connector. Update `.pre-commit-config.yaml` to
use this version of detect-secrets.