PP-9545: agreement search status accepts both upper and lower case values
Use the commons `AgreementStatus` enum to accept agreement search
params. Internally (in the database, in events) this is represented as
the default string version of an enum which default to caps. Public API
will be passing in parsed strings from users which will be lowercase.

Using a case insensitive match on the search param will allow Ledger to
return the expected data but will throw a reasonable error if values
that aren't present in the enum (and therefore can't have been written
to the database) are passed in.