PP-9292 Send raw event message to SNS
Previously, we were deserialising the event message coming from
SQS queue and then re-serialising it to send it on to SNS.

We weren't re-serialising the event in an identical way, and were
serialising the `event_details` as an escaped string rather than a
nested JSON object.

To avoid these deserialisation mistakes, get the raw string contained in
the original SQS message to send to SNS. The intention is that ledger
does not change the message, and this will ensure this.