Add Exemption object to Ledger Specification
With this commit, we are starting to introduce the exemption object in the
Ledger specification so that it will be added to the TransactionView.

If an exemption was requested, the new object will be set with the property
`requested` set to true. If we know that an exemption was not requested, then
the new object will be set with the property `requested` set to false.

Further information in the JIRA ticket below.

https://payments-platform.atlassian.net/browse/PP-13332