Introduce type and outcome result to Exemption object
With this change, we are updating the Exemption object by adding a new
property `type` to represent the type of the exemption request, and a new
object `outcome` containing the property `result` to represent the outcome of
the exemption request.

The new property `type` is set to `corporate` or `optimised`, or is not set
when we don't have this information.

The new property `result`, within `outcome`, will be set to:
- `honoured`, when the exemption request is successful
- `out of scope`, when the request is not correct, e.g. for a corporate
  exemption but with a normal credit card
- `rejected`, when the exemption request is not successful

Further information in Jira.

https://payments-platform.atlassian.net/browse/PP-13332