Test all the various exemption scenarios
With this change, we are testing all the exemption scenarios, considering a
transaction for which the outcome is known or not, whether we know that the
exemption was requested or we know that it was not requested, and whether the
exemption was of type corporate, otherwise we can deduce that it was an
optimised exemption request.

Further information in Jira.

https://payments-platform.atlassian.net/browse/PP-13332