Update the Exemption creation flow to match Connector
With this change, we are moving around branches of the flow within the
Exemption creation method in order to closely match the method
ChargeResponse.Exemption buildExemptionResponseObject
(Exemption3ds exemption3ds, Exemption3dsType exemption3dsRequested) in
Connector.

This will make it easier to eventually hoist the code into Pay Java commons
and use it from there.

Further information in the PR for Ledger[1], and in the PR for Connector[2].

[1]
https://github.com/alphagov/pay-ledger/pull/3651#pullrequestreview-2468638568

[2]
https://github.com/alphagov/pay-connector/pull/5636#discussion_r1848519600