PP-13622 Fix incorrect negation for won disputes in CSV files
With this change, we are fixing a bug for which disputed amounts are shown as
negative numbers in the CSV files.

The logic will now consider also the state of the dispute and present the
value as positive when the dispute has been won by the service.

Further information in Jira.

https://payments-platform.atlassian.net/browse/PP-13622